requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
openpyxl>=3.1
psycopg[binary]>=3.1
whitenoise>=6.6
gunicorn>=21.2
